requirements.txt
langchain-core
langchain_community
langchain_google_genai
langchain_openai
google-api-python-client
colorama
python-dotenv
bs4
html2text
requests
notion-client
pydantic
